Page collecting system
Added a way to collect "pages" using an array and a collision detecting function (similar to walls.) When the player walks over a page, that page is removed from the array and map while the "pagescollected" integer is increased by 1.

I also added a display in the top left corner that show the player how many pages they have collected.