Main menu and Map Layout
Made a simple map layout with the pages spread throughout the rooms.

Also made states for when the game is running or is in main menu. With a simple main menu added as well.

Also added a small page counter in the top left corner that increases when the player collects a page