Win-Lose Conditions
Added a "monster". The monster starts in the other side of the maps and activates a chase when the player collects their second page. The monster will then walk towards the player until it either catches up or the player collects the fifth page.

If the player collects all five pages, they recieve a game over screen saying they won. If the monster catches them, they recieve the same screen but with "you were caught".

Also added a main menu that the user starts on when opening the game. The player also has the ability to either return to the main menu or play again when reaching the "game over" screen.