Classes
Added and incorporated two classes. One less important for the player, and one for the monster. Also added an objects for each of the classes with the added in values that were used before.

Classes was supposed to be a key addition in the code however we failed to consider them when writing.